chore(web): add header directory
