+++ packages/web/src/app/(authenticated)/inbox/components/message/item/header/label-type.ts
