+++ packages/web/src/app/(authenticated)/inbox/components/message/item/header/sentiment.tsx
@@ -1,0 +1,3 @@
+export default function Sentiment({ sentiment }: { sentiment: string }) {
+  return <span className='text-xl'>{sentiment}</span>;
+}
