+++ packages/web/src/app/(authenticated)/inbox/components/message/item/header/time.tsx
@@ -1,0 +1,5 @@
+import dayjs from 'dayjs';
+
+export default function Time({ send_at }: { send_at: string }) {
+  return <span className=''>{dayjs(send_at).format('H:mm')}</span>;
+}
