--- packages/web/src/app/(authenticated)/inbox/components/message/item/sentiment.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function Sentiment() {
-  return <span className='text-xl'>😀</span>;
-}
