--- packages/web/src/app/(authenticated)/inbox/components/message/item/time.tsx
@@ -1,6 +1,0 @@
-import dayjs from 'dayjs';
-import React from 'react';
-
-export default function Time({ send_at }: { send_at: string }) {
-  return <span className=''>{dayjs(send_at).format('HH:mm')}</span>;
-}
